% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_functions.R
\name{align_sequences}
\alias{align_sequences}
\title{Align sequences}
\usage{
align_sequences(filename, gene = NULL, uniprot = NULL,
  shouldtextplot = FALSE, textplotcolumns = NULL)
}
\arguments{
\item{filename}{file with sequence information, see Details for needed columns}

\item{gene}{genename (or other name-identifier) for the alignment.}

\item{uniprot}{Uniprot ID (or other ID-identifier) for the alignment.}

\item{shouldtextplot}{logical, should a table with sequence info be plotted below the alignment?}

\item{textplotcolumns}{columns to include if printing a table below the alignment.}
}
\description{
Align sequence to a full length protein (exact matches).
}
\details{
The input file needs to have one row per peptide/protein
    (full length sequence to align as the top row) and the columns\cr
    - Name (identifier to be written above the sequence)\cr
    - Sequence1 (1 in column name also if only one sequence per row)\cr
    - Sequence2 (only if mosaic, not necessary if only one sequence per row)\cr
    - Color (color for the sequence bar)\cr
    - Include (values 1 or 0, which sequences in input file should be included in plot?)

If printing table underneath extra columns may be needed in the input file
    depending on what information you want to print.
}
