% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{AminoAcids}
\alias{AminoAcids}
\title{Data frame with amino acid names}
\format{
An object of class \code{\link[base:data.frame]{"data.frame"}}.
}
\usage{
data(AminoAcids)
}
\description{
Data frame with full amino acid names, shorthand and one letter
}
\keyword{datasets}
