% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_functions.R
\name{ap_aaindex}
\alias{ap_aaindex}
\title{Find amino acid indices}
\usage{
ap_aaindex(x, fasta_list = NULL, sequence_list = NULL)
}
\arguments{
\item{x}{List with at least one element, see Deatils for naming and content.}

\item{fasta_list}{A list of sequences in fasta format. Either this or sequence_list should be provided.}

\item{sequence_list}{A list of only sequences. Either this or fasta_list should be provided.}
}
\value{
Updated input x with new columns in the BEADS element for the start and stopp indices of each tested isoform.
}
\description{
Align query sequnces to full length sequences to find start and stopp indices of sequences.
}
\details{
NB! The sequences have to be identical matches to consequtive stretches
    in the full length sequence you are aligning to. A future update will include partial matches.

    The x list needs to include at least the element

    BEADS = Beads info with identifiable rownames and with at least the columns:

    "Sequence" with one letter abriviation for the aminoa acids,

    "Uniprot" withthe uniprot ID the sequence originates from

    If a list of fasta sequences is provided (\code{fasta_list}), it should be the output of
    \code{\link[seqinr:read.fasta]{read.fasta()}} in the package \code{seqinr}.

    If a list of only sequenes is provided (\code{sequence_list}),
    each element should correspond to a Uniprot ID, isoform specific if applicable,
     and contain a consequitive string of one letter amino acid abbreviations.
}
