% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{ap_binary}
\alias{ap_binary}
\title{Binary}
\usage{
ap_binary(x, cutoffs)
}
\arguments{
\item{x}{List of scoring values with two levels per element: level one = assay data sets ;
level two =  bead subsets (e.g. wih and w/o controls).
It is recommended to use to element Scoring in the output from \link[rappp]{ap_scoring}).}

\item{cutoffs}{data.frame with at least one column named score with the desired cutoffs to use,
and rownames you want to have as identifier for each cutoff.
It is recommended is to use the Cutoff_key element in the output from \link[rappp]{ap_scoring}).}
}
\value{
List with binary data.frames
}
\description{
Create binary matrices based on scored Autoimmunity profiling data.
}
\details{
The input values will be binned into binary data, consisting of 0 and 1.

The input values should be scoring values, and structured as a list
(preferably the output from function ap_scoring()), even if only one data set is used
(see examples in \link[rappp]{ap_mads}).
}
