% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_2_diff_input.R
\name{ap_binary2}
\alias{ap_binary2}
\title{Binary}
\usage{
ap_binary2(x, cutoffs)
}
\arguments{
\item{x}{List with at least one element, see Deatils for naming and content.
It is recommended to use to element SCORE in the output from \code{\link[rappp:ap_scoring2]{ap_scoring2()}}.}

\item{cutoffs}{data.frame with at least one column named score with the desired cutoffs to use,
and rownames you want to have as identifier for each cutoff.
It is recommended to use the COKEY element in the output from \code{\link[rappp:ap_scoring2]{ap_scoring2()}}.}
}
\value{
Updated input x with the new list element

    BINARY = list with one data.frame per cutoff
}
\description{
Create binary matrices based on scored Autoimmunity profiling data.
}
\details{
The input values will be binned into binary data, consisting of 0 and 1.

The x list needs to include at least the element:

    SCORE = scored data,
}
