% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{ap_cutoff_selection}
\alias{ap_cutoff_selection}
\title{Cutoff selection}
\usage{
ap_cutoff_selection(x, cutoffs, slope_cutoff = -0.5, offset = 0.1, bw = 0.1)
}
\arguments{
\item{x}{List of scoring values with two levels per element: level one = assay data sets ;
level two =  bead subsets (e.g. wih and w/o controls).
It is recommended to use to element Scoring in the output from \link[rappp]{ap_scoring}).}

\item{cutoffs}{data.frame with at least two columns:

    - One column named score with the desired cutoffs to use

    - One column named xmad with the corresponding MAD cutoff values

It is recommended is to use the Cutoff_key element in the output from \link[rappp]{ap_scoring}).}

\item{slope_cutoff}{Arbitrary slope cutoff value. Can be chosen freely.}

\item{offset}{Offset used to prevent script from finding the peak (as slope = 0 there).}

\item{bw}{Bandwidth for density funciton, default set to 0.1.}
}
\value{
List with two main elements

    [[1]] Density output used for cutoff selection, with same structure as input list.

    [[2]] Calculated antigen specific cutoffs, with same structure as input list.
}
\description{
Select cutoff based on the slope of the density of scores per antigen.
}
\details{
A cutoff will be selected for each antigen based on the
distribution of the scores for the antigen.
The algorithm will search for a local min nearest the highest
peak in a density plot using bandwidth=0.1.

The input values should be scoring values, and structured as a list
(preferably the output from function ap_scoring()), even if only one data set is used
(see examples in \link[rappp]{ap_mads}).
}
