% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_2_diff_input.R
\name{ap_cutoff_selection2}
\alias{ap_cutoff_selection2}
\title{Cutoff selection}
\usage{
ap_cutoff_selection2(x, slope_cutoff = -0.5, offset = 0.1, bw = 0.1,
  check.names = FALSE)
}
\arguments{
\item{x}{List with at least two elements, see Details for naming and content.
It is recommended to use the output from \code{\link[rappp:ap_scoring2]{ap_scoring2()}}.}

\item{slope_cutoff}{Arbitrary slope cutoff value. Can be chosen freely.}

\item{offset}{Offset used to prevent script from finding the peak (as slope = 0 there).}

\item{bw}{Bandwidth for density funciton, default set to 0.1.}

\item{check.names}{logical, altered default from \code{\link[base:data.frame]{data.frame()}}.}
}
\value{
Updated input x with the new list elements

    DENS = Density output used for cutoff selection,

    ANTIGEN_CUTOFFS_CONT = Calculated antigen specific cutoffs, continues values,

    ANTIGEN_CUTOFFS = Calculated antigen specific cutoffs, translated into the descrete cutoff steps,

    BINARY_CO = Binary table based on the antigen specific cutoffs.
}
\description{
Select cutoff based on the slope of the density of scores per antigen.
}
\details{
A cutoff will be selected for each antigen based on the
distribution of the scores for the antigen. The algorithm will search for a
local min nearest the highest peak in a density plot using bandwidth=0.1.

The x list needs to include at least the element:

    SCORE = scored data,

    COKEY = Cutoff key as data.frame with cutoff values, scores and colors.
}
