% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_2_diff_input.R
\name{ap_cutoffs2}
\alias{ap_cutoffs2}
\title{Cutoff key}
\usage{
ap_cutoffs2(MADlimits = seq(0, 70, 5))
}
\arguments{
\item{MADlimits}{vector of MADs values used as boundaries for binning (≥MADs).}
}
\value{
data.frame with three columns:

   [,1] MADs cutoff value

   [,2] Corresponding score value

   [,3] Corresponding color using the Zissou1 palette in \code{\link[wesanderson]{wes_palette}}
}
\description{
Create a cutoff key for scoring of Autoimmunity Profiling data.
}
\details{
The input values will be binned into discrete bins (scores).
}
