% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_functions.R
\name{ap_cutoffs2image}
\alias{ap_cutoffs2image}
\title{Cutoff key image}
\usage{
ap_cutoffs2image(cutoffkey = NULL, MADlimits = NULL)
}
\arguments{
\item{cutoffkey}{table matching output from \code{\link[rappp:ap_cutoffs2]{ap_cutoffs2()}},
recommended input if \code{\link[rappp:ap_norm2]{ap_norm2()}} has been used.}

\item{MADlimits}{vector of MADs values used as boundaries for binning (≥MADs), eg. seq(0,70,5).
Not used if cutoffkey is provided.}
}
\value{
If MADlimits is provided a data.frame with three columns will be returned:

   [,1] MADs cutoff value

   [,2] Corresponding score value

   [,3] Corresponding color using the Zissou1 palette in \code{\link[wesanderson]{wes_palette}}
}
\description{
Create a cutoff key for scoring of Autoimmunity Profiling data and
also produce an image.
Uses \code{\link[rappp:ap_cutoffs2]{ap_cutoffs2()}}
}
