% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_functions.R
\name{ap_excel}
\alias{ap_excel}
\title{Export data from Autoimmunity Profiling analysis to excel}
\usage{
ap_excel(
  x,
  elements = c("MFI", "MADS", "SCORE", "BINARY_CO", "REACTIVITIES", "FISHER",
    "DIFFERENCES", "ANTIGEN_CUTOFFS", "CUTOFF_KEY", "ANTIGENS", "SAMPLES", "COUNT"),
  filename = "DataOutput.xlsx",
  shouldround = TRUE,
  row.names = TRUE,
  save_rdata = TRUE,
  ...
)
}
\arguments{
\item{x}{list with at least the elements to export, see Deatils for more information and exceptions to the rule.}

\item{elements}{character vector of names of the list elements to export.
The sheets will be ordered in the same order as the vector.
Specific MAD-cutoffs are not included in the default output but can be added, eg. "25xMAD".}

\item{filename}{string with filename and desired path, end with .xlsx.}

\item{shouldround}{logical, if TRUE MFI values are rounded to integers and MADs to two decimals.}

\item{row.names}{logical. If TRUE, the row names of the data frames are included in the Excel file worksheets.
Deafult altered from \code{\link[WriteXLS:WriteXLS]{WriteXLS()}}.}

\item{save_rdata}{logical. If TRUE, an RData-file is saved, containg a list matching the Excel-file content.}

\item{...}{arguments passed to \code{\link[WriteXLS:WriteXLS]{WriteXLS()}}.}
}
\description{
Exports chosen list elemnts to sheets in an Excel-file.
Default elements are based on output from \code{\link[rappp:ap_norm2]{ap_norm2()}}, and
 \code{\link[rappp:ap_reactsummary2]{ap_reactsummary2()}} or \code{\link[rappp:ap_agresults]{ap_agresults()}}
}
\details{
The x list needs to include at least the elements specified under \code{elements}.
  It is recommended to append the output from \code{\link[rappp:ap_reactsummary2]{ap_reactsummary2()}} or
  \code{\link[rappp:ap_agresults]{ap_agresults()}} to the output from \code{\link[rappp:ap_norm2]{ap_norm2()}}
  and use the combined list as function input.

  Exceptions for input element names:\cr
  - If an element is named BEADS in the input data, its name will be changed to ANTIGENS.
  Therefore, ANTIGENS is a default element to export while BEADS is not.\cr
  - Sums and frequencies at the cutoffs assigned by \code{\link[rappp:ap_cutoff_selection2]{ap_cutoff_selection2()}}
  will be combined into a new element called REACTIVITIES
 if the output from \code{\link[rappp:ap_reactsummary2]{ap_reactsummary2()}} or
 \code{\link[rappp:ap_agresults]{ap_agresults()}} is included in the output
 Therefore, REACTIVITIES is a default element to export although it is not present in the input data.\cr
 - If samplegorups were provided in \code{\link[rappp:ap_reactsummary2]{ap_reactsummary2()}} the
 resulting Fisher's exact test p-values and frequency differences will be extracted for the
 the cutoffs assigned by \code{\link[rappp:ap_cutoff_selection2]{ap_cutoff_selection2()}} and formatted
 into the new elements FISHER and DIFFERENCES.
 Therefore, FISHER and DIFFERENCES are default elements to export although they are not present in the input data.\cr

 Line breaks (backslash n) in row and column names will be changed to underscore ("_").

  If other list structures are used, it is most likely more convenient to just use
  \code{\link[WriteXLS:WriteXLS]{WriteXLS()}}, which this function is built on.
}
