% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_functions.R
\name{ap_igx}
\alias{ap_igx}
\title{Loading control}
\usage{
ap_igx(x, IgX_bead, IgType = "G", IgX_cutoff = 5000, cosfac = c(3,
  -3), shouldplot = T, filename = "anti-humanIgX.pdf", width = 10,
  height = 6, useDingbats = F)
}
\arguments{
\item{x}{List with at least three elements, see Deatils for naming and content.}

\item{IgX_bead}{Column index for empty bead.}

\item{IgType}{Which Imunoglobulin is measured, default is G.}

\item{IgX_cutoff}{MFI cutoff value for filtering.}

\item{cosfac}{Median absolute deviation multipliers in vector c(upper, lower),
for drawing lines and detecting potential outliers.}

\item{shouldplot}{Logical, should a plot be made?}

\item{filename}{String with filename and desired path, end with .pdf}

\item{width, height}{Width and height for pdf, see \code{\link[grDevices:pdf]{pdf()}}.}

\item{useDingbats}{Logical. Default is \code{FALSE}, compared to in default \code{\link[grDevices:pdf]{pdf()}}.}
}
\value{
Updated input x with relevant filtering info and a pdf with plot (if \code{shouldplot=T}).
}
\description{
Filter samples with low MFI for the anti-human IgX bead.
}
\details{
The x list needs to include at least the elements:

    MFI = assay mfi,

    SAMPLES = Sample info. See below for required columns.

    FILTERINFO = Vector with info on which filter steps has been done.

The SAMPLES element needs at least the columns:

    "sample_name" with sample names, preferably LIMS-IDs, where
    replicates (named with one of pool|rep|mix|commercial)
    and blanks (named with one of empty|blank|buffer) are also stated,

    "AssayNum" with assay number (vector with 1s if only one assay, support for up to 5 assys in one plot),

    "AssayWell" with Well IDs in the assay plate, e.g A01, B01 etc.,

    "tube_label" with alternative sample names, eg. from collaborator,
}
