% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_2_diff_input.R
\name{ap_mads2}
\alias{ap_mads2}
\title{MAD normalization}
\usage{
ap_mads2(x, constant = 1, na.rm = TRUE, low = FALSE, high = FALSE,
  check.names = FALSE)
}
\arguments{
\item{x}{list with at least two elements, see Details for naming and content.}

\item{constant}{constant for \code{\link[stats:mad]{mad()}} function,
default is 1 (compared to 1.4826 in base function).}

\item{na.rm}{logical, indicating whether NA values should be stripped
before the computation proceeds. Altered default from
\code{\link[stats:median]{median()}} and \code{\link[stats:mad]{mad()}}.}

\item{low}{if TRUE, compute the ‘lo-median’, i.e., for even sample size, do not average
the two middle values, but take the smaller one.(From \code{\link[stats:mad]{mad()}}).}

\item{high}{if TRUE, compute the ‘hi-median’, i.e., take the larger of the two middle values
for even sample size.(From \code{\link[stats:mad]{mad()}}).}

\item{check.names}{logical, altered default from \code{\link[base:data.frame]{data.frame()}}.}
}
\value{
Updated input x with the new list element

    MADS = assay MADs.
}
\description{
Sample based normalization to number of Median Absolute Deviations (MADs)
from the median for Autoimmunity profiling data.
}
\details{
The input values will be normalized per sample to the number of MADs from the median
using the algorithm MADs = (MFI - median )/MAD, where MAD is calculated using \code{mad(constant=1)}.

The x list needs to include at least the elements:

    MFI = assay mfi,

    BEADS = Beads info, if any should be excluded then these should be annotated in a column called "Filtered".
    Any beads with no text (ie. "" or NA) or "NegControl" in such column will be included in the transformation.
}
