% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_functions.R
\name{ap_negbeads}
\alias{ap_negbeads}
\title{Overview of signals in relation to neg control beads}
\usage{
ap_negbeads(
  x,
  shouldpdf = TRUE,
  filename = "neg-control-beads.pdf",
  width = 15,
  height = 10,
  useDingbats = FALSE
)
}
\arguments{
\item{x}{List with at least four elements, see Deatils for naming and content.}

\item{shouldpdf}{Logical, should it plot to pdf?}

\item{filename}{String with filename and desired path, end with .pdf}

\item{width, height}{Width and height for pdf, see \code{\link[grDevices:pdf]{pdf()}}.}

\item{useDingbats}{Logical. Default is \code{FALSE}, compared to in default \code{\link[grDevices:pdf]{pdf()}}.}
}
\description{
Plot overview of singals in relation to the neg control beads (Empty, His6ABP, Neutravidin or Biotin).
Based on output from Autoimmunity Profiling scoring function \code{\link[rappp:ap_scoring2]{ap_scoring2()}}.
}
\details{
The x list needs to include at least the element

    MFI = assay mfi, column names of negative control columns should include empty|bare|blank|his6abp|hisabp|neutravidin,

    SCORE = scored data, column names of negative control columns should include empty|bare|blank|his6abp|hisabp|neutravidin,

    CUTOFF_KEY = Cutoff key as data.frame with cutoff values, scores and colors.

    BEADS = Beads info, including a column called "Filtered" with the value "NegControl" for the negative control beads.
    Any beads with no text (ie. "" or NA) or "NegControl" in such column will be included.

    SAMPLES = Sample info, if any should be excluded then these should be annotated in a column called "Filtered".
    Any samples with no text (ie. "" or NA) in such column will be included.

Note: The function plots to a layout containing three areas.
}
