% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{ap_norm}
\alias{ap_norm}
\title{Cutoff selection}
\usage{
ap_norm(x, MADlimits = seq(0, 70, 5), ...)
}
\arguments{
\item{x}{List of MFI values with two levels per element: level one = assay data sets ;
level two =  bead subsets (e.g. wih and w/o controls)}

\item{MADlimits}{vector of MADs values used as boundaries for binning (≥MADs).}

\item{...}{See respective functions for details: \link[rappp]{ap_mads}, \link[rappp]{ap_scoring},
\link[rappp]{ap_binary}, \link[rappp]{ap_cutoff_selection}}
}
\value{
List with 7 main elements

    [[1]] Input MFI

    [[2]] MADs

    [[3]] Scores

    [[4]] Binary

    [[5]] Antigen specific cutoffs selected based on density slope.

    [[6]] Density information

    [[7]] Cutoff translation key
}
\description{
Wrapper function for full AUtoimmunity Profiling data transformations.
}
\details{
The input values should be MFI values, and structured as a list,
even if only one data set is used (see examples).
}
