% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_2_diff_input.R
\name{ap_norm2}
\alias{ap_norm2}
\title{Full AP data transformation}
\usage{
ap_norm2(x, MADlimits = seq(0, 70, 5), ...)
}
\arguments{
\item{x}{List with at least two elements, see Deatils for naming and content.}

\item{MADlimits}{vector of MADs values used as boundaries for binning (≥MADs).}

\item{...}{See respective functions for details:
\code{\link[rappp:ap_mads2]{ap_mads2()}}, \code{\link[rappp:ap_scoring2]{ap_scoring2()}},
\code{\link[rappp:ap_binary2]{ap_binary2()}}, \code{\link[rappp:ap_cutoff_selection2]{ap_cutoff_selection2()}}.}
}
\value{
Updated input x with the new list elements

    MADs = assay MADs,

    COKEY = Cutoff key as data.frame with cutoff values, scores and colors,

    SCORE = scored data,

    BINARY = list with one data.frame per cutoff,

    DENS = Density output used for cutoff selection,

    AGCO = Calculated antigen specific cutoffs, translated into the descrete cutoff steps,

    AGCO_CONT = Calculated antigen specific cutoffs, continues values.
}
\description{
Wrapper function for full Autoimmunity Profiling data transformations.
}
\details{
The x list needs to include at least the elements:

    MFI = assay mfi,

    BEADS = Beads info (Filtered column with information about filtering),
}
