% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_functions.R
\name{ap_overview}
\alias{ap_overview}
\title{Signal overview}
\usage{
ap_overview(
  x,
  grepneg = "his6abp|hisabp|empty|bare|biotin|neutravidin|neg_",
  greppos = "anti-h|hIg|ebna",
  includeFilter = TRUE,
  Aglabels = "Gene_agID",
  shouldpdf = TRUE,
  filename = "Signal_overview.pdf",
  width = 25,
  height = 15,
  useDingbats = FALSE,
  ...
)
}
\arguments{
\item{x}{List with at least one elements, see Deatils for naming and content.}

\item{grepneg}{Regular expression for color of negative control beads.}

\item{greppos}{Regular expression for color of positive control beads.}

\item{includeFilter}{Logical, should antigen boxes be colored based on filtering?}

\item{Aglabels}{Column in x$BEADS with matching names as columns in x$MFI. Only used if includeFilter is TRUE.}

\item{shouldpdf}{Logical, should it plot to pdf?}

\item{filename}{String with filename and desired path, end with .pdf}

\item{width, height}{Width and height for pdf, see \code{\link[grDevices:pdf]{pdf()}}.}

\item{useDingbats}{Logical. Default is \code{FALSE}, compared to in default \code{\link[grDevices:pdf]{pdf()}}.}

\item{...}{Further arguments passed to \code{\link[graphics:boxplot]{boxplot()}}.}
}
\description{
Boxplots of signals per antigen and sample in different orders.
}
\details{
The x list needs to include at least the element
    MFI = assay mfi.

    BEADS = Beads info. Only needed if includeFilter is TRUE.

    Data points with the value NA or 0 will be set to 1 for the plotting to allow for
    logarithmic scale without filtering any beads or samples.

Note: The function plots to a layout containing three areas.
}
