% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_2_diff_input.R
\name{ap_reactsummary2}
\alias{ap_reactsummary2}
\title{Calculate reactivity frequencies}
\usage{
ap_reactsummary2(x, samplegroups = NULL)
}
\arguments{
\item{x}{List with at least two elements, see Deatils for naming and content.}

\item{samplegroups}{factor vector of groupings. Only samples with an assigned level are included in plots.
If left as \code{NULL} (default), the all non-filtered, if filetring done otherwise all, will be assigned "Sample".}
}
\value{
Updated input x with the new list element

    REACTSUM_AG = number of reactive samples per antigen and sample group,

    REACTFREQ_AG = reactivity frequency per antigen and sample group,

    REACTSUM_SAMP = number of reactive antigens per sample,

    REACTFREQ_SAMP = reactivity frequency per sample,
}
\description{
Create binary matrices based on scored Autoimmunity profiling data.
}
\details{
The x list needs to include at least the elements:

    BINARY = list with one data.frame per cutoff

    BINARY_CO = Binary table based on the antigen specific cutoffs.
}
