% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_functions.R
\name{ap_rep}
\alias{ap_rep}
\title{Check replicate samples}
\usage{
ap_rep(x, iter = 500, filename = "replicates.pdf", width = 12,
  height = 12, useDingbats = F)
}
\arguments{
\item{x}{List with at least two elements, see Deatils for naming and content.}

\item{iter}{How many times  random samples should be iterated.}

\item{filename}{String with filename and desired path, end with .pdf}

\item{width, height}{Width and height for pdf, see \code{\link[grDevices:pdf]{pdf()}}.}

\item{useDingbats}{Logical. Default is \code{FALSE}, compared to in default \code{\link[grDevices:pdf]{pdf()}}.}
}
\description{
various plots in one PDF to assess the replicates and overall reproducibility.
}
\details{
The x list needs to include at least the element
    MFI = assay mfi,

    SAMPLES = Sample info. See below for required columns.

The SAMPLES element needs at least the columns:

    "sample_name" with sample names, preferably LIMS-IDs, where
    replicates (named with one of pool|rep|mix|commercial)

    "AssayNum" with assay number (vector with 1s if only one assay),
}
