% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{ap_scoring}
\alias{ap_scoring}
\title{Scoring}
\usage{
ap_scoring(x, rightmost.closed = FALSE, left.open = FALSE,
  check.names = FALSE, ...)
}
\arguments{
\item{x}{List of MADs values with two levels per element: level one = assay data sets ;
level two =  bead subsets (e.g. wih and w/o controls)}

\item{rightmost.closed, left.open}{logical, see \link[base]{findInterval} for details.}

\item{check.names}{logical, see \link[base]{data.frame} for details}

\item{...}{Further arguments passed do \link[base]{findInterval}}
}
\value{
List of scored data, with same structure as input list.
}
\description{
Binning of MADs values in Autoimmunity Profiling.
}
\details{
The input values will be binned into discrete bins (scores).

The input values should be MADs values, and structured as a list
(preferably the output from functino ap_mads()), even if only one data set is used, see examples.
}
