% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_2_diff_input.R
\name{ap_scoring2}
\alias{ap_scoring2}
\title{Scoring}
\usage{
ap_scoring2(x, MADlimits = seq(0, 70, 5), rightmost.closed = FALSE,
  left.open = FALSE, check.names = FALSE, ...)
}
\arguments{
\item{x}{List with at least one elements, see Deatils for naming and content.
It is recommended to use the the output from \code{\link[rappp:ap_mads2]{ap_mads2()}}.}

\item{MADlimits}{vector of MADs values used as boundaries for binning (≥MADs).}

\item{rightmost.closed, left.open}{logical, see \code{\link[base:findInterval]{findInterval()}} for details.}

\item{check.names}{logical, see \code{\link[base:data.frame]{data.frame()}} for details}

\item{...}{Further arguments passed do \code{\link[base:findInterval]{findInterval()}}}
}
\value{
Updated input x with the new list elements

    COKEY = Cutoff key as data.frame with cutoff values, scores and colors

    SCORE = scored data
}
\description{
Binning of MADs values in Autoimmunity Profiling.
}
\details{
The input values will be binned into discrete bins (scores).

The x list needs to include at least the element:

    MADs = assay MADs,
}
