% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap_textplot.R
\name{ap_textplot}
\alias{ap_textplot}
\title{Display text information in a graphics plot.}
\usage{
ap_textplot(object, halign = "center", valign = "center", cex, ...)
}
\description{
This function displays text output in a graphics window.
It is the equivalent of 'print' except that the output is displayed as a plot.
}
\details{
Based on \code{\link[gplots:textplot]{textplot()}} with slightly altered code.
Altered code found onlne (https://gist.github.com/johncolby/1482973).
Additional alteration done by CH, including making working methods, see code.

The input should be one of classes \code{matrix}, \code{data.frame}, \code{vectors} longer than 1,
single \code{character} strings, single \code{integers} or single \code{numeric} values.
}
