% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bead_count.R
\name{bead_count}
\alias{bead_count}
\title{Bead count}
\usage{
bead_count(x)
}
\arguments{
\item{bead_count_file}{A dataframe with bead count. Each row is one well, each column is a bead.}
}
\value{
Plot with bead count per sample and per bead
}
\description{
Bead count
}
\details{
Investigate if you have enough beads in each well of each beadID
}
\section{Warnings}{

This is a warning
}

\examples{
bead_count(df)
}
