% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bead_count.R
\name{bead_count_base}
\alias{bead_count_base}
\title{Bead count with ggplot}
\usage{
bead_count_base(
  bead_count_df,
  bead_count_limit = 32,
  path_for_output = "./",
  date_in_file_name = T,
  ...
)
}
\arguments{
\item{bead_count_df}{A dataframe with bead count. Each row is a well, each column is a bead.}

\item{bead_count_limit}{How many beads do you need? Can be vector if you want several lines.}

\item{path_for_output}{Path for output}

\item{date_in_file_name}{Yes or No}
}
\value{
Plot with bead count per sample and per bead
}
\description{
Investigate if you have enough beads in each well of each beadID
}
\section{Warnings}{

This is a warning
}

\examples{
bead_count_base(df)
}
