% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_plots.R
\name{beeswarm_ggplot}
\alias{beeswarm_ggplot}
\title{Beeswarm plots for SBA data using ggplot}
\usage{
beeswarm_ggplot(data, filename, sample_info, bead_info, grouping_var,
  sample_name_var, xlab)
}
\arguments{
\item{data}{A df with data from an SBA experiment. May be transformed or not.}

\item{filename}{A string with the desired name of the output \code{.pdf}.}

\item{sample_info}{A df with the sample info.}

\item{bead_info}{A df with the bead info.}

\item{grouping_var}{A string with the name of the column in \code{sample_info} containing the desired grouping variable.}

\item{sample_name_var}{A string with the name of the column in \code{sample_info} containing the desired sample identifier.}

\item{xlab}{A character vector containing the desired grouping labels to be printed.}
}
\description{
Makes beeswarm (jitter) plots for all analytes (beads) in the input object \code{data}.
Grouping is performed according to the \code{grouping_var} column in the df \code{sample_info}.
}
