% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_small_functions.R
\name{cv}
\alias{cv}
\title{Coefficient of Variation}
\usage{
cv(x, format = "percent", digits = 2, ...)
}
\arguments{
\item{x}{Vector or matrix with values.}

\item{format}{If the output should be as "percent" (default) or "decimal".}

\item{digits}{Integer indicating the number of decimal places (see \link[stats]{round})}

\item{...}{Further arguments passed do \link[stats]{mean} and \link[stats]{sd}}
}
\value{
The CV of the input values.
}
\description{
This function computes the Coefficient of Variation (CV) of the values in x.
}
\details{
The argument na.rm can be included. Default is FALSE in both \link[stats]{mean} and \link[stats]{sd}.
If set to TRUE, then missing values are removed before computation proceeds.
}
