% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_small_functions.R
\name{layout384}
\alias{layout384}
\title{Print a 384-well plate layout}
\usage{
layout384(x, n = 1:384)
}
\arguments{
\item{x}{values to print in cells/wells}

\item{n}{indicies for where the values should be printed, default is all wells.}
}
\description{
Create a 384-well plate layout from vector format.
}
