% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_small_functions.R
\name{layout96}
\alias{layout96}
\title{Print a 96-well plate layout}
\usage{
layout96(x, n = 1:96)
}
\arguments{
\item{x}{values to print in cells/wells}

\item{n}{indicies for where the values should be printed, default is all wells.}
}
\description{
Create a 96-well plate layout from vector format.
}
