% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_functions.R
\name{make_peptides}
\alias{make_peptides}
\title{Make peptides}
\usage{
make_peptides(sequence, lengths = 15:20, min_overlap = 2,
  min_shift = 1)
}
\arguments{
\item{sequence}{string with amino acid sequence to make peptides from,
see Examples for how to use with the Clipboard.}

\item{lengths}{number of amino acids per peptide, can be a vector of lenghts.}

\item{min_overlap}{the shortest allowed overlap between peptides}

\item{min_shift}{the shortest allowed shift between peptides}
}
\value{
A list with one element per design
}
\description{
Get possible peptide designs across different lengths and overlaps.
}
\examples{
## The input sequence can easily be copied from wherever,
    like Excel, LIMS etc and used directly in the function:
## on PC
make_peptides(readClipboard())

## on MAC
make_peptides(t(read.table(pipe("pbpaste"), stringsAsFactors = FALSE)))

}
