% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{MockSBA_processed}
\alias{MockSBA_processed}
\title{A precessed mock SBA list}
\format{An object of class \code{\link[base:list]{"list"}} with 22 elements:
\describe{
  \item{MFI}{A data.frame with raw MFI values, samples as rows and beads as columns.}
  \item{COUNT}{A data.frame with bead count, samples as rows and beads as columns.}
  \item{SAMPLES}{A data.frame with sample information, samples as rows.}
  \item{BEADS}{A data.frame with bead information, beads as rows.}
  \item{CT}{A data.frame with raw MFI values from the coupling efficiency test, replicates as rows and beads as columns.}

  \item{FILTERINFO}{A string vector, starting will NULL but is expanded depending on performed filtering steps.}
  \item{MADS}{A data.frame with MADs values, samples as rows and beads as columns.
        Output from \code{\link[rappp:ap_mads2]{"ap_mads2"}} and \code{\link[rappp:ap_norm2]{"ap_norm2"}}.}
  \item{CUTOFF_KEY}{A data.frame with the translation between MADs, score and colors.
        Output from \code{\link[rappp:ap_scoring2]{"ap_scoring2"}} and \code{\link[rappp:ap_norm2]{"ap_norm2"}}.}
  \item{SCORE}{A data.frame with score values, samples as rows and beads as columns.
        Output from \code{\link[rappp:ap_scoring2]{"ap_scoring2"}} and \code{\link[rappp:ap_norm2]{"ap_norm2"}}.}
  \item{BINARY}{A list of data.frames with binary values, one data.frame per cutoff in CUTOFF_KEY, samples as rows and beads as columns.
        Output from \code{\link[rappp:ap_binary2]{"ap_binary2"}} and \code{\link[rappp:ap_norm2]{"ap_norm2"}}.}
  \item{DENS}{A list of outputs from \code{\link[stats:density]{"density"}}, one element per bead.}
  \item{ANTIGEN_CUTOFFS_CONT}{A numeric vector with the x-values selected
        by \code{\link[rappp:ap_cutoff_selection2]{"ap_cutoff_selection2"}} and \code{\link[rappp:ap_norm2]{"ap_norm2"}} from DENS.}
  \item{ANTIGEN_CUTOFFS}{A tibble the selected cutoff score and corresponding MADs for each antigen.
        Output from \code{\link[rappp:ap_cutoff_selection2]{"ap_cutoff_selection2"}} and \code{\link[rappp:ap_norm2]{"ap_norm2"}}.}
  \item{BINARY_CO}{A data.frame with binary values, using the antigen specific cutoffs in ANTIGEN_CUTOFFS, samples as rows and beads as columns.
        Output from \code{\link[rappp:ap_cutoff_selection2]{"ap_cutoff_selection2"}} and \code{\link[rappp:ap_norm2]{"ap_norm2"}}.}

  \item{SAMPLEGROUPS}{A data.frame with used sample grouping in \code{\link[rappp:ap_reactsummary2]{"ap_reactsummary2"}} and
        \code{\link[rappp:ap_agresults]{"ap_agresults"}}. The output from these functions will be in a separate list,
        which was appended to the outout from \code{\link[rappp:ap_norm2]{"ap_norm2"}} in this example dataset.}
  \item{REACTSUM_AG}{A data.frame with reactivity sums per antigen (column) and cutoff-grouping combination (rows)
        from \code{\link[rappp:ap_reactsummary2]{"ap_reactsummary2"}} and \code{\link[rappp:ap_agresults]{"ap_agresults"}}.
        The output from these functions will be in a separate list,
        which was appended to the outout from \code{\link[rappp:ap_norm2]{"ap_norm2"}} in this example dataset.}
  \item{REACTFREQ_AG}{A data.frame with reactivity frequencies per antigen (column) and cutoff-grouping combination (rows)
        from \code{\link[rappp:ap_reactsummary2]{"ap_reactsummary2"}} and \code{\link[rappp:ap_agresults]{"ap_agresults"}}.
        The output from these functions will be in a separate list,
        which was appended to the outout from \code{\link[rappp:ap_norm2]{"ap_norm2"}} in this example dataset.}
  \item{REACTSUM_SAMP}{A data.frame with reactivity sums per sample (row) and cutoff (column)
        from \code{\link[rappp:ap_reactsummary2]{"ap_reactsummary2"}} and \code{\link[rappp:ap_agresults]{"ap_agresults"}}.
        The output from these functions will be in a separate list,
        which was appended to the outout from \code{\link[rappp:ap_norm2]{"ap_norm2"}} in this example dataset.}
  \item{REACTFREQ_SAMP}{A data.frame with reactivity frequencies per sample (row) and cutoff (column)
        from \code{\link[rappp:ap_reactsummary2]{"ap_reactsummary2"}} and \code{\link[rappp:ap_agresults]{"ap_agresults"}}.
        The output from these functions will be in a separate list,
        which was appended to the outout from \code{\link[rappp:ap_norm2]{"ap_norm2"}} in this example dataset.}

  \item{COMPARISONS}{A matrix with group comparisons (column) made if sample grouping includes at least two levels
        in \code{\link[rappp:ap_reactsummary2]{"ap_reactsummary2"}} and \code{\link[rappp:ap_agresults]{"ap_agresults"}}.
        The output from these functions will be in a separate list,
        which was appended to the outout from \code{\link[rappp:ap_norm2]{"ap_norm2"}} in this example dataset.}
  \item{FISHER_P}{A list of data.frames with p-values from Fisher's exact test between groups in COMPARISONS.
        One data.frame per group comparison, with antigens as columns and cutoff as rows.
        Outout from \code{\link[rappp:ap_reactsummary2]{"ap_reactsummary2"}} and \code{\link[rappp:ap_agresults]{"ap_agresults"}}.
        The output from these functions will be in a separate list,
        which was appended to the outout from \code{\link[rappp:ap_norm2]{"ap_norm2"}} in this example dataset.}
  \item{FREQ_DIFF}{A list of data.frames with frequency differences between groups in COMPARISONS.
        One data.frame per group comparison, with antigens as columns and cutoff as rows.
        Outout from \code{\link[rappp:ap_reactsummary2]{"ap_reactsummary2"}} and \code{\link[rappp:ap_agresults]{"ap_agresults"}}.
        The output from these functions will be in a separate list,
        which was appended to the outout from \code{\link[rappp:ap_norm2]{"ap_norm2"}} in this example dataset.}
}}
\usage{
data(MockSBA_processed)
}
\description{
A list with a mock autoimmunity SBA data set, including processed data.
}
\keyword{datasets}
