% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_small_functions.R
\name{pairs2}
\alias{pairs2}
\title{Scatterplot Matrices}
\usage{
pairs2(
  x,
  labels,
  panel = points,
  ...,
  horInd = 1:nc,
  verInd = 1:nc,
  lower.panel = panel,
  upper.panel = panel,
  diag.panel = NULL,
  text.panel = textPanel,
  label.pos = 0.5 + has.diag/3,
  line.main = 3,
  cex.labels = NULL,
  font.labels = 1,
  row1attop = TRUE,
  gap = 1,
  log = ""
)
}
\arguments{
\item{x}{the coordinates of points given as numeric columns of a matrix or data frame.
Logical and factor columns are converted to numeric in the same way that
\code{\link[base:data.matrix]{data.matrix()}} does.}

\item{labels}{the names of the variables.}

\item{panel}{\code{function(x, y, ...)} which is used to plot the contents of each panel of the display.}

\item{...}{arguments to be passed to or from methods.
Also, graphical parameters (\code{\link[graphics:par]{par()}}) can be given as can arguments to
\code{plot} such as \code{main}.
\code{par("oma")} will be set appropriately unless specified.}

\item{horInd, verInd}{The (numerical) indices of the variables to be plotted on the horizontal and vertical axes respectively.}

\item{lower.panel, upper.panel}{separate panel functions (or \code{NULL}) to be used below and above the diagonal respectively.}

\item{diag.panel}{optional \code{function(x, ...)} to be applied on the diagonals.}

\item{text.panel}{optional \code{function(x, y, labels, cex, font, ...)} to be applied on the diagonals.}

\item{label.pos}{y position of labels in the text panel.}

\item{line.main}{if \code{main} is specified, \code{line.main} gives the \code{line argument} to
\code{\link[graphics:mtext]{mtext()}}
which draws the title. You may want to specify \code{oma} when changing \code{line.main}.}

\item{cex.labels, font.labels}{graphics parameters for the text panel.}

\item{row1attop}{logical. Should the layout be matrix-like with row 1 at the top, or graph-like with row 1 at the bottom?
The latter (non default) leads to a basically symmetric scatterplot matrix.}

\item{gap}{distance between subplots, in margin lines.}

\item{log}{a character string indicating if logarithmic axes are to be used, see
\code{\link[graphics:plot.default]{plot.default()}}
or a numeric vector of indices specifying the indices of those
variables where logarithmic axes should be used for both x and y.
\code{log = "xy"} specifies logarithmic axes for all variables.}
}
\description{
A matrix of scatterplots is produced.
Alternative version of \code{\link[graphics:pairs]{pairs()}} (default S3 method) where all axes are on the bottom and left sides.
Argument information copied from \code{\link[graphics:pairs]{pairs()}}.
}
\details{
Please see \code{\link[graphics:pairs]{pairs()}}
}
