% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Textplot.R
\name{Textplot}
\alias{Textplot}
\alias{Textplot.default}
\alias{Textplot.matrix}
\alias{Textplot.data.frame}
\alias{Textplot.character}
\title{Display text information in a graphics plot.}
\usage{
Textplot(object, ...)

\method{Textplot}{default}(object, ...)

\method{Textplot}{matrix}(object, hPlotAlign = c("center", "left",
  "right"), vPlotAlign = c("center", "top", "bottom"), cex, mar = c(1,
  1, 4, 1) + 0.1, cmar = 1, rmar = 1, showRowNames = TRUE,
  showColumnNames = TRUE, hCellAlign = c("center", "left", "right"),
  vCellAlign = c("center", "top", "bottom"), dataColor = par("col"),
  rowNameColor = par("col"), columnNameColor = par("col"), ...)

\method{Textplot}{data.frame}(object, ...)

\method{Textplot}{character}(object, hPlotAlign = c("center", "left",
  "right"), vPlotAlign = c("center", "top", "bottom"), cex, mar = c(0,
  0, 3, 0) + 0.1, ...)
}
\arguments{
\item{object}{Object to be displayed.  If character of length > 1, then elements
are concatenated with a newline delimiter.}

\item{...}{Optional arguments passed to the text plotting command.}

\item{hPlotAlign}{Character with alignment of the entire object within the plotting region
in the x direction, one of "center", "left", or "right".  For character \samp{object} with multiple lines of text
this parameter is also used to align the text within the plot
(i.e. "center" will center the object within the plotting region and also center 2
or more lines of text within the plot itself)}

\item{vPlotAlign}{Character with alignment of the entire object within the plotting region in the y direction,
one of "center", "left", or "right".}

\item{cex}{Numeric of length 1 with the text character size, see \link{par} for details.
If missing, the code will attempt to use the largest value which allows the entire plot region to be filled.}

\item{mar}{Figure margins, see the documentation for \link{par}}

\item{cmar}{Numeric with the space between columns, in fractions of the size of the letter 'M'.  Typically > 1}

\item{rmar}{Numeric with the space between columns, in fractions of the size of the letter 'M'.  Typically > 1}

\item{showRowNames}{Logical indicating whether row names will be displayed (for matrix, data.frame, or vector objects)}

\item{showColumnNames}{Logical indicating whether column names will be displayed (for matrix, data.frame, or vector objects)}

\item{hCellAlign}{Character with the horizontal location of elements within cell
(for matrix, data.frame, or vector objects).  One of "center", "left", or "right".}

\item{vCellAlign}{Character with the vertical location of elements within cell
(for matrix, data.frame, or vector objects).  One of "center", "top", or "bottom".}

\item{dataColor}{Colors for data elements. If a single value is provided,
 all data elements will be the same color. If a matrix matching the dimensions of the data is provided,
each data element will receive the specified color.}

\item{rowNameColor}{Colors for row names. May be specified as a scalar or a vector of the same row length as object.}

\item{columnNameColor}{Colors for column names. May be specified as a scalar or a vector of the same column length as object.}
}
\value{
Numeric with the auto-sized cex value if cex is missing, otherwise the original cex value.
}
\description{
This function displays text output in a graphics window.
There is an option to display the text using the largest font that will fit in the plotting region.
For matrixes, data.frames and vectors a specialized textplot function
is available that plots each of the cells individually in a way that is visually
appealing (maintains the table-like/grid-like structure of the data).
If present, row and column labels will be displayed in a bold font.
}
\details{
This function was modified from the PerformanceAnalytics version by Peter Carl and Brian G. Peterson (brian@braverock.com).
Assistance was provided by John Colby in this post:
\url{http://stackoverflow.com/questions/8523944/left-justify-a-column-using-textplot-gplots-or-performanceanalytics}

Modifications:
\itemize{
            \item More descriptive variables names, more in-line comments,
            more detailed/clear parameter documentation, more concise code
            \item Bug fix and easier to use/understand cex parameter.
            If cex is specified then what you specify is what you get.
            Previously a specified cex was adjusted by par(cex) by the text() function, which is confusing.
            \item This function sets par(cex = 1) and then passes the cex parameter
            (or computed cex in the auto-size case) to text(), so caller simply decides
            what cex is desired or omits cex to auto-size.
            \item Bug fix to overlapping text when left aligning matrix/data.frame
            (see StackOverflow post linked above)
            \item Eliminate vadj and hadj in favor of vCellAlign and hCellAlign and renamed
            valign/halign to vPlotAlign and hPlotAlign.
            Caller decides "center", "left", "right", "top", "bottom" for plot and cells and it all works out.
            \item Using a small hack, textplot.character will horizontally align text within
            the plot region and within the plot itself.
            Previously the text was only aligned within the plot region so it appeared
            as if the alignment instruction was ignored.
}

Color has NOT been tested.
Don't deal with the TAB character.
}
\author{
Suraj Gupta (suraj at wingedfootcapital.com)
}
