% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_functions.R
\name{tsne_perp}
\alias{tsne_perp}
\title{tSNE plots with different perplexities}
\usage{
tsne_perp(
  z,
  perp = c(2, 5, 10, 50),
  sqrt = TRUE,
  iterations = 1000,
  groups,
  names,
  legend = T,
  legendname = NULL,
  main = NULL,
  filename = "t-SNE_perplexities.pdf",
  height = 16,
  useDingbats = F
)
}
\arguments{
\item{z}{the data as a data.frame}

\item{perp}{a vector of values to test as perplexities, eg. c(2,5,10,50).}

\item{sqrt}{logical, should sqrt(Nsamples) be added to perplexities?
Potentially the lowest needed perplexity.}

\item{iterations}{number of iterations per tSNE}

\item{groups}{grouping for colors and lines}

\item{names}{text to display as point labels}

\item{legend}{TRUE or FALSE}

\item{legendname}{Title of legend}

\item{main}{Title of page}

\item{filename}{String with filename and desired path, end with .pdf}

\item{height}{height for pdf, see \code{\link[grDevices:pdf]{pdf()}}.}

\item{useDingbats}{Logical. Default is \code{FALSE}, compared to in default \code{\link[grDevices:pdf]{pdf()}}.}
}
\description{
tSNE plots with different perplexities and designs.
}
