% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_functions.R
\name{align_sequences}
\alias{align_sequences}
\title{Align sequences}
\usage{
align_sequences(inputfile, outputfile = "SequenceAlignment.pdf",
  gene = NULL, uniprot = NULL, shouldtextplot = FALSE,
  textplotcolumns = NULL)
}
\arguments{
\item{inputfile}{file with sequence information, see Details for needed formatting and columns.}

\item{outputfile}{filename for the output file.}

\item{gene}{string with genename (or other name-identifier) for the alignment.}

\item{uniprot}{string with Uniprot ID for the alignment.}

\item{shouldtextplot}{logical, should a table with sequence info be plotted below the alignment?}

\item{textplotcolumns}{columns to include if printing a table below the alignment.}

\item{shouldpdf}{Logical, should it plot to pdf?}
}
\description{
Visualize how sequences align to a full length protein (exact matches).
}
\details{
NB! The sequences have to be identical matches to consequtive stretches
    in the full length sequence you are aligning to.

    The input file needs to be tab-delimited, have one row per peptide/protein and the columns\cr
    - Name (identifier to be written above the sequence)\cr
    - Type ("Full_length" for sequence to align against,
            "Peptide" for peptides (sequence will be printed on the side),
            any other label can be used but will not be considered in the function)
    - Sequence1 (1 in column name also if only one sequence per row)\cr
    - Sequence2 (only if mosaic, not necessary if only one sequence per row)\cr
    - Color (color for the sequence bar)\cr
    - Include (values 1 or 0, which sequences in input file should be included in plot?)

If printing table underneath extra columns may be needed in the input file
    depending on what information you want to print.

Note: The function plots to a layout containing up to two areas.
}
