% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_functions.R
\name{ap_agresults}
\alias{ap_agresults}
\title{Reactivity result plots}
\usage{
ap_agresults(x, samplegroups = NULL, groupcolors = 1:6,
  agtoplot = NULL, filename = "AntigenResults.pdf", height = 15,
  useDingbats = F)
}
\arguments{
\item{x}{List with at least nine elements, see Deatils for naming and content.}

\item{samplegroups}{factor vector of groupings. Only samples with an assigned level are included in plots.
If left as \code{NULL} (default), the all non-filtered, if filetring done otherwise all, will be assigned "Sample".}

\item{groupcolors}{colors for each group in samplegroups.}

\item{agtoplot}{Indices for which antigens to plot, default is all.
Character vector with column names of what to plot also ok.}

\item{filename}{String with filename and desired path, end with .pdf}

\item{height}{Width and height for pdf, see \code{\link[grDevices:pdf]{pdf()}}.}

\item{useDingbats}{Logical. Default is \code{FALSE}, compared to in default \code{\link[grDevices:pdf]{pdf()}}.}
}
\value{
A list with the elements

    REACTSUM_AG = number of reactive samples per antigen and sample group,

    REACTFREQ_AG = reactivity frequency per antigen and sample group,

    REACTSUM_SAMP = number of reactive antigens per sample,

    REACTFREQ_SAMP = reactivity frequency per sample,
}
\description{
Plot beeswarm, density and frequency plots for each antigen.
Based on output from Autoimmunity Profiling wrapper function \code{\link[rappp:ap_norm2]{ap_norm2()}}.
}
\details{
The x list needs to include at least the element

    MADs = assay MADs,

    SCORE = scored data

    BINARY = list with one data.frame per cutoff,

    BINARY_CO = binary table based on antigen specific cutoffs.

    ANTIGEN_CUTOFFS = calculated antigen specific cutoffs, translated into the descrete cutoff steps,

    CUTOFF_KEY = cutoff key as data.frame with cutoff values, scores and colors.

    SAMPLES = sample info. Including column "sample_name" with sample names, preferably LIMS-IDs, where
    replicates (named with one of pool|rep|mix|commercial)
    and blanks (named with one of empty|blank|buffer) are also stated,

    BEADS = beads info, if any should be excluded then these should be annotated in a column called "Filtered".
    Any beads with no text (ie. "") will be included in the transformation.

    DENS = Density output used for cutoff selection,
}
