% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_2_diff_input.R
\name{ap_binary2}
\alias{ap_binary2}
\title{Binary}
\usage{
ap_binary2(x, check.names = FALSE)
}
\arguments{
\item{x}{List with at least two elements, see Details for naming and content.
It is recommended to use the output from \code{\link[rappp:ap_scoring2]{ap_scoring2()}}.}

\item{check.names}{logical, altered default from \code{\link[base:data.frame]{data.frame()}}.}
}
\value{
Updated input x with the new list element

    BINARY = list with one data.frame per cutoff
}
\description{
Create binary matrices based on scored Autoimmunity profiling data.
}
\details{
The input values will be binned into binary data, consisting of 0 and 1.

The x list needs to include at least the elements:

    SCORE = scored data,

    CUTOFF_KEY = Cutoff key as data.frame with cutoff values, scores and colors.
}
