% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_functions.R
\name{ap_count}
\alias{ap_count}
\title{Bead count}
\usage{
ap_count(x, labels = "Gene_HPRR", protein = "GeneShort",
  agID = "PrEST", samp_co = 32, bead_flag = 32, bead_filter = 16,
  N_filter = 0, shouldplot = T, filename = "bead_count.pdf",
  width = 12, height = 10, useDingbats = F)
}
\arguments{
\item{x}{List with at least four elements, see Deatils for naming and content.}

\item{labels}{Column name in BEADS with antigen names to be used in pdf.}

\item{protein}{Column name in BEADS with short protein or gene name.}

\item{agID}{Column name in BEADS with antigen identifier, eg. PrEST ID or product number.}

\item{samp_co}{Cutoff for filtering samples with low median count.}

\item{bead_flag}{Cutoff for flagging beads with low counts.}

\item{bead_filter}{Cutoff for filtering beads with low counts.}

\item{N_filter}{Accepted number of samples with low count per bead ID.}

\item{shouldplot}{Logical, should a plot be made?}

\item{filename}{String with filename and desired path, end with .pdf}

\item{width, height}{Width and height for pdf, see \code{\link[grDevices:pdf]{pdf()}}.}

\item{useDingbats}{Logical. Default is \code{FALSE}, compared to in default \code{\link[grDevices:pdf]{pdf()}}.}
}
\value{
Updated input x with relevant filtering and/or flagging info and a pdf with plots (if \code{shouldplot=T}).
}
\description{
Filter and/or flag samples and beads with low bead count.
}
\details{
The x list needs to include at least the elements:

    COUNT = bead count,

    BEADS = Beads info. See below for required columns.

    SAMPLES = Sample info. See below for required columns.

    FILTERINFO = Vector with info on which filter steps has been done.

The BEADS element needs at least the columns:

    "Plate" with coupling plate number(s),

    "BeadID" with bead ID number,

    separate columns with full desired labels, protein/gene names and product ID numbers.

The SAMPLES element needs at least the columns:

    "Sample" with sample names, preferably LIMS-IDs, where
    replicates (named with one of pool|rep|mix|commercial)
    and blanks (named with one of empty|blank|buffer) are also stated,

    "AssayWell" with Well IDs in the assay plate, e.g A01, B01 etc.,

    "tube_label" with alternative sample names, eg. from collaborator,

    "WashCol", a numerical vector with groupings for during-run-washes in Luminex.
    Eg. c(rep(1, 96), rep(2, 96), rep(1, 96), rep(2, 96)) if washes every 96th well in a 384-well plate.
    Colors boxes accordingly. There are 6 colors available.
}
