% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_functions.R
\name{ap_ct}
\alias{ap_ct}
\title{Coupling efficency test}
\usage{
ap_ct(x, empty_bead, empty_co_multiple = 3, shouldplot = T,
  filename = "coupling_efficiency.pdf", width = 30, height = 6,
  useDingbats = F, ...)
}
\arguments{
\item{x}{List with at least three elements, see Deatils for naming and content.}

\item{empty_bead}{Column index for empty bead.}

\item{empty_co_multiple}{Number of sd above empty for cutoff.}

\item{shouldplot}{Logical, should a plot be made?}

\item{filename}{String with filename and desired path, end with .pdf}

\item{width, height}{Width and height for pdf, see \code{\link[grDevices:pdf]{pdf()}}.}

\item{useDingbats}{Logical. Default is \code{FALSE}, compared to in default \code{\link[grDevices:pdf]{pdf()}}.}

\item{...}{Further arguments passed to \code{\link[base:mean]{mean()}} and \code{\link[stats:sd]{sd()}}.}
}
\value{
Updated input x with relevant filtering info and a pdf with plot (if shouldplot=T).
}
\description{
Flags beads with signal similar to empty bead in coupling test, produces plot if wanted.
}
\details{
The x list needs to include at least the elements
    CT = coupling test mfi,

    BEADS = Beads info (including Type-column with PrEST for PrESTs),

    FILTERINFO = Vector with info on which filter steps has been done.
}
