% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{ap_ct}
\alias{ap_ct}
\title{Coupling efficency test}
\usage{
ap_ct(x, empty_bead, empty_co_multiple = 3, shouldplot = T,
  filename = "coupling_efficiency.pdf", width = 30, height = 6,
  useDingbats = F, ...)
}
\arguments{
\item{x}{List with at least three elements:
CT = coupling test mfi,
BEADS = Beads info (including Type-column with PrEST for PrESTs),
FILTERINFO = Vector with info on whichfilter steps has been done.}

\item{empty_bead}{Column index for empty bead.}

\item{empty_co_multiple}{Number of sd above empty for cutoff.}

\item{shouldplot}{Logical, should a plot be made?}

\item{filename}{String with filename and desired path, end with .pdf}

\item{width, height}{Width and height for pdf, see \link[grDevices]{pdf}.}

\item{useDingbats}{Logical. Default is \code{FALSE}, compared to in default \link[grDevices]{pdf}.}

\item{...}{Further arguments passed to \link[stats]{mean}, \link[stats]{sd}, and \link[grDevices]{pdf}.}
}
\description{
Flags beads with signal similar to empty bead in coupling test, produces plot if wanted.
}
