% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_2_diff_input.R
\name{ap_cutoff_selection2}
\alias{ap_cutoff_selection2}
\title{Cutoff selection}
\usage{
ap_cutoff_selection2(x, cutoffs, slope_cutoff = -0.5, offset = 0.1,
  bw = 0.1)
}
\arguments{
\item{x}{List with at least one element, see Deatils for naming and content.
It is recommended to use to element Scoring in the output from \code{\link[rappp:ap_scoring2]{ap_scoring2()}}.}

\item{cutoffs}{data.frame with at least one column named score with the desired cutoffs to use,
and rownames you want to have as identifier for each cutoff.
It is recommended to use the COKEY element in the output from \code{\link[rappp:ap_scoring2]{ap_scoring2()}}.}

\item{slope_cutoff}{Arbitrary slope cutoff value. Can be chosen freely.}

\item{offset}{Offset used to prevent script from finding the peak (as slope = 0 there).}

\item{bw}{Bandwidth for density funciton, default set to 0.1.}
}
\value{
Updated input x with the new list elements

    DENS = Density output used for cutoff selection.

    AGCO = Calculated antigen specific cutoffs, translated into the descrete cutoff steps.

    AGCO_CONT = Calculated antigen specific cutoffs, continues values.
}
\description{
Select cutoff based on the slope of the density of scores per antigen.
}
\details{
A cutoff will be selected for each antigen based on the
distribution of the scores for the antigen. The algorithm will search for a
local min nearest the highest peak in a density plot using bandwidth=0.1.

The x list needs to include at least the element:

    SCORE = scored data,
}
