% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_functions.R
\name{ap_igx}
\alias{ap_igx}
\title{Loading control}
\usage{
ap_igx(
  x,
  IgX_bead,
  IgType = "G",
  IgX_cutoff = 5000,
  cosfac = c(3, -3),
  internal_sampID = "sample_name",
  external_sampID = "tube_label",
  shouldplot = TRUE,
  shouldpdf = TRUE,
  filename = "anti-humanIgX.pdf",
  width = 12,
  height = 6,
  useDingbats = FALSE,
  ...
)
}
\arguments{
\item{x}{List with at least three elements, see Deatils for naming and content.}

\item{IgX_bead}{Column index for empty bead.}

\item{IgType}{Which Imunoglobulin is measured, default is G.}

\item{IgX_cutoff}{MFI cutoff value for filtering.}

\item{cosfac}{Median absolute deviation multipliers in vector c(upper, lower),
for drawing lines and detecting potential outliers.}

\item{internal_sampID}{Column name in SAMPLES with internal sample IDs, such as from LIMS.
Replicates (named with one of pool|rep|mix|commercial, not case sensitive)
and blanks (named with one of empty|blank|buffer, not case sensitive) must be stated in this column.}

\item{external_sampID}{Column name in SAMPLES with externaö sample IDs, such as given from the collaborator or user.}

\item{shouldplot}{Logical, should a plot be made?}

\item{shouldpdf}{Logical, should it plot to pdf?}

\item{filename}{String with filename and desired path, end with .pdf}

\item{width, height}{Width and height for pdf, see \code{\link[grDevices:pdf]{pdf()}}.}

\item{useDingbats}{Logical. Default is \code{FALSE}, compared to in default \code{\link[grDevices:pdf]{pdf()}}.}

\item{...}{Arguments are passed to base plot function.}
}
\value{
Updated input x with relevant filtering info and a pdf
    with plot (if \code{shouldplot=TRUE} and \code{shouldpdf=TRUE}).
}
\description{
Filter samples with low MFI for the anti-human IgX bead.
}
\details{
The x list needs to include at least the elements:

    MFI = assay mfi,

    SAMPLES = Sample info. See below for required columns.

    FILTERINFO = Vector with info on which filter steps has been done.

The SAMPLES element needs at least the columns:
    "AssayNum" with assay number (vector with 1s if only one assay, support for up to 5 assys in one plot),

    "AssayWell" with Well IDs in the assay plate, e.g A01, B01 etc.,

Note: The function plots to a layout containing seven areas.
}
