% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{ap_igx}
\alias{ap_igx}
\title{Loading control (FAR FROM DONE)}
\usage{
ap_igx(x, IgX_bead, IgType = "G", IgX_cutoff = 5000, cosfac = c(3,
  -3), shouldplot = T, filename, width = 10, height = 6,
  useDingbats = F)
}
\arguments{
\item{x}{List with at least three elements:
MFI = assay mfi,
SAMPLES = Sample info. See Details for required columns.
BEADS = Beads info (including Type-column with PrEST for PrESTs),
FILTERINFO = Vector with info on whichfilter steps has been done.}

\item{IgX_bead}{Column index for empty bead.}

\item{IgType}{Which Imunoglobulin is measured, default is G.}

\item{IgX_cutoff}{MFI cutoff value for filtering.}

\item{cosfac}{Median absolute deviation multipliers in vector c(upper, lower),
for drawing lines and detecting potential outliers.}

\item{shouldplot}{Logical, should a plot be made?}

\item{filename}{String with filename and desired path, end with .pdf}

\item{width, height}{Width and height for pdf, see \link[grDevices]{pdf}.}

\item{useDingbats}{Logical. Default is \code{FALSE}, compared to in default \link[grDevices]{pdf}.}
}
\description{
Filter samples with low MFI for the anti-human IgX bead.
}
\details{
The SAMPLES element needs at least the columns:

    "Sample" with sample names, preferably LIMS-IDs, where
    replicates (named with one of pool|rep|mix|commercial)
    and blanks (named with one of empty|blank|buffer) are also stated,

    "AssayNum" with assay number (vector with 1s if only one assay, support for up to 5 assys in one plot),

    "Well384" with Well IDs, e.g A01, B01 etc.,

    "tube_label" with alternative sample names, eg. from collaborator,
}
