% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_2_diff_input.R
\name{ap_mads2}
\alias{ap_mads2}
\title{MAD normalization}
\usage{
ap_mads2(x, constant = 1, na.rm = TRUE, ...)
}
\arguments{
\item{x}{list with at least two elements, see Deatils for naming and content.}

\item{constant}{constant for \code{\link[stats:mad]{mad()}} function, default is 1 (compared to 1.4826 in base function).}

\item{na.rm}{logical, indicating whether NA values should be stripped before the computation proceeds. Altered default from
\code{\link[stats:median]{median()}} and \code{\link[stats:mad]{mad()}}.}

\item{...}{Further arguments passed to \code{\link[stats:mad]{mad()}}}
}
\value{
Updated input x with the new list element

    MADs = assay MADs.
}
\description{
Sample based normalization to number of Median Absolute Deviations (MADs)
from the median for Autoimmunity profiling data.
}
\details{
The input values will be normalized per sample to the number of MADs from the median
using the algorithm MADs = (MFI - median )/MAD, where MAD is calculated using \code{mad(constant=1)}.

The x list needs to include at least the elements:

    MFI = assay mfi,

    BEADS = Beads info, if any should be excluded then these should be annotated in a column called "Filtered".
    Any beads with no text (ie. "") or "NegControl" in such column will be included in the transformation.
}
