% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_functions.R
\name{ap_negbeads}
\alias{ap_negbeads}
\title{Overview of signals in relation to neg control beads}
\usage{
ap_negbeads(x, filename = "neg-control-beads.pdf", width = 15,
  height = 10, useDingbats = F)
}
\arguments{
\item{x}{List with at least four elements, see Deatils for naming and content.}

\item{filename}{String with filename and desired path, end with .pdf}

\item{width, height}{Width and height for pdf, see \code{\link[grDevices:pdf]{pdf()}}.}

\item{useDingbats}{Logical. Default is \code{FALSE}, compared to in default \code{\link[grDevices:pdf]{pdf()}}.}
}
\description{
Plot overview of singals in relation to the neg control beads.
Based on output from Autoimmunity Profiling scoring function \code{\link[rappp:ap_scoring2]{ap_scoring2()}}.
}
\details{
The x list needs to include at least the element
    MFI = assay mfi,

    SCORE = scored data,

    COKEY = Cutoff key as data.frame with cutoff values, scores and colors.

    BEADS = Beads info, if any should be excluded then these should be annotated in a column called "Filtered".
    Any beads with no text (ie. "") or "NegControl" in such column will be included in the transformation.
}
