% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_2_diff_input.R
\name{ap_norm2}
\alias{ap_norm2}
\title{Full AP data transformation}
\usage{
ap_norm2(x, MADlimits = seq(0, 70, 5), na.rm = TRUE,
  check.names = FALSE, mad_constant = 1, mad_low = FALSE,
  mad_high = FALSE, score_rightmost.closed = FALSE,
  score_left.open = FALSE, score_all.inside = FALSE,
  coselect_slope_cutoff = -0.5, coselect_offset = 0.1,
  coselect_bw = 0.1)
}
\arguments{
\item{x}{List with at least three elements, see Details for naming and content.}

\item{MADlimits}{vector of MADs values used as boundaries for binning (≥MADs).}

\item{na.rm}{logical, indicating whether NA values should be stripped
before the computation proceeds. Altered default from
\code{\link[stats:median]{median()}} and \code{\link[stats:mad]{mad()}}.}

\item{check.names}{logical, altered default from \code{\link[base:data.frame]{data.frame()}}.}

\item{mad_constant}{constant for \code{\link[stats:mad]{mad()}} function,
default is 1 (compared to 1.4826 in base function).}

\item{mad_low}{if TRUE, compute the ‘lo-median’, i.e., for even sample size, do not average
the two middle values, but take the smaller one.(From \code{\link[stats:mad]{mad()}}).}

\item{mad_high}{if TRUE, compute the ‘hi-median’, i.e., take the larger of the two middle values
for even sample size.(From \code{\link[stats:mad]{mad()}}).}

\item{score_rightmost.closed, score_left.open, score_all.inside}{logical,
see \code{\link[base:findInterval]{findInterval()}} for details.
Defaults result in scores for MADS ≥ cutoff, and any value below the lowest cutoff gets score 0.}

\item{coselect_slope_cutoff}{Arbitrary slope cutoff value. Can be chosen freely.}

\item{coselect_offset}{Offset used to prevent script from finding the peak (as slope = 0 there).}

\item{coselect_bw}{Bandwidth for density funciton, default set to 0.1.}

\item{samplegroups}{factor vector of groupings. Only samples with an assigned level are included in plots.
If left as \code{NULL} (default), the all non-filtered, if filetring done otherwise all, will be assigned "Sample".}
}
\value{
Updated input x with the new list elements

    MADs = assay MADs,

    COKEY = Cutoff key as data.frame with cutoff values, scores and colors,

    SCORE = scored data,

    BINARY = list with one data.frame per cutoff,

    DENS = Density output used for cutoff selection,

    AGCO = Calculated antigen specific cutoffs, translated into the descrete cutoff steps,

    AGCO_CONT = Calculated antigen specific cutoffs, continues values.

    REACTSUM_AG = number of reactive samples per antigen and sample group,

    REACTFREQ_AG = reactivity frequency per antigen and sample group,

    REACTSUM_SAMP = number of reactive antigens per sample,

    REACTFREQ_SAMP = reactivity frequency per sample.
}
\description{
Wrapper function for full Autoimmunity Profiling data transformations.
}
\details{
Arguments starting with mad_ are specific for \code{\link[rappp:ap_mads2]{ap_mads2()}},
score_ for \code{\link[rappp:ap_scoring2]{ap_scoring2()}},
and coselect_ for \code{\link[rappp:ap_cutoff_selection2]{ap_cutoff_selection2()}}.
These arguments are seldom altered.

The x list needs to include at least the elements:

    MFI = assay mfi,

    BEADS = Beads info (Filtered column with information about filtering),

    SAMPLES = Sample info, if any should be excluded then these should be annotated in a column called "Filtered".
    Any beads with no text (ie. "") in such column will be included.
}
