% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_functions.R
\name{ap_overview}
\alias{ap_overview}
\title{Signal overview}
\usage{
ap_overview(x, filename = "Signal_overview.pdf", width = 20,
  height = 15, useDingbats = F, ...)
}
\arguments{
\item{x}{List with at least one elements, see Deatils for naming and content.}

\item{filename}{String with filename and desired path, end with .pdf}

\item{width, height}{Width and height for pdf, see \code{\link[grDevices:pdf]{pdf()}}.}

\item{useDingbats}{Logical. Default is \code{FALSE}, compared to in default \code{\link[grDevices:pdf]{pdf()}}.}

\item{...}{Further arguments passed to \code{\link[graphics:boxplot]{boxplot()}}.}
}
\description{
Boxplots of signals per antigen and sample in different orders.
}
\details{
The x list needs to include at least the element
    MFI = assay mfi,
}
