% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_2_diff_input.R
\name{ap_reactsummary2}
\alias{ap_reactsummary2}
\title{Calculate reactivity frequencies}
\usage{
ap_reactsummary2(x, samplegroups = NULL, check.names = FALSE)
}
\arguments{
\item{x}{List with at least three elements, see Details for naming and content.}

\item{samplegroups}{factor vector of groupings. Only samples with an assigned level are included in plots.
If left as \code{NULL} (default), the all non-filtered, if filetring done otherwise all, will be assigned "Sample".}

\item{check.names}{logical, altered default from \code{\link[base:data.frame]{data.frame()}}.}
}
\value{
Updated input x with the new list element

    REACTSUM_AG = number of reactive samples per antigen and sample group,

    REACTFREQ_AG = reactivity frequency per antigen and sample group,

    REACTSUM_SAMP = number of reactive antigens per sample,

    REACTFREQ_SAMP = reactivity frequency per sample,
}
\description{
Create binary matrices based on scored Autoimmunity profiling data.
}
\details{
The x list needs to include at least the elements:

    SAMPLES = Sample info, if any should be excluded then these should be annotated in a column called "Filtered".
    Any beads with no text (ie. "") in such column will be included.

    BINARY = list with one data.frame per cutoff

    BINARY_CO = Binary table based on the antigen specific cutoffs.
}
