% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_2_diff_input.R
\name{ap_scoring2}
\alias{ap_scoring2}
\title{Scoring}
\usage{
ap_scoring2(
  x,
  MADlimits = seq(0, 70, 5),
  rightmost.closed = FALSE,
  left.open = FALSE,
  all.inside = FALSE,
  check.names = FALSE
)
}
\arguments{
\item{x}{List with at least one element, see Details for naming and content.
It is recommended to use the the output from \code{\link[rappp:ap_mads2]{ap_mads2()}}.}

\item{MADlimits}{vector of MADs values used as boundaries for binning.}

\item{rightmost.closed, left.open, all.inside}{logical, see \code{\link[base:findInterval]{findInterval()}} for details.
Defaults result in scores for MADS ≥ cutoff, and any value below the lowest cutoff gets score 0.}

\item{check.names}{logical, altered default from \code{\link[base:data.frame]{data.frame()}}.}
}
\value{
Updated input x with the new list elements

    CUTOFF_KEY = Cutoff key as data.frame with cutoff values, scores and colors

    SCORE = scored data
}
\description{
Binning of MADs values in Autoimmunity Profiling.
}
\details{
The input values will be binned into discrete bins (scores).

The x list needs to include at least the element:

    MADS = assay MADs,
}
