% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_functions.R
\name{ap_summary}
\alias{ap_summary}
\title{Analysis summary}
\usage{
ap_summary(x)
}
\arguments{
\item{x}{List with at least two elements, see Deatils for naming and content.}
}
\description{
Summarizes number of filtered/flagged beads/samples, amino acid lenghts and protein representation.
}
\details{
The x list needs to include at least the elements

    SAMPLES = Sample info. Including column "sample_name" with LIMS-IDs.

    BEADS = Beads info. Including columns:

    "Type" with info about type of content on bead, at least including "PrEST" for PrESTs,

    "PrEST.seq..aa." with amino acid sequences,

    "Filtered" with filtering annotation, e.g. from other ap_-functions

    "Flagged" with filtering annotation, e.g. from other ap_-functions
}
