% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap_textplot.R
\name{ap_textplot}
\alias{ap_textplot}
\title{Display text information in a graphics plot.}
\usage{
ap_textplot(object, halign = "center", valign = "center", cex, ...)
}
\description{
This function displays text output in a graphics window.
It is the equivalent of 'print' except that the output is displayed as a plot.
Based \code{\link[graphics:pairs]{pairs()}} with altered deafults
}
\details{
Please see \code{\link[graphics:pairs]{pairs()}}
}
